// src/shared/enums/activity-action-type.enum.ts
export enum ActivityLogActionType {
   TICKET_CREATION = "ticket_creation",
   TICKET_DATA_UPDATE = "ticket_data_update",
   TICKET_DATA_ADD = "ticket_data_add",
   TICKET_DATA_REMOVE = "ticket_data_remove",
   OTHER_FEES_REMOVE = "other_fees_remove",
   FORM_START = "form_start",
   FORM_DATA_UPDATE = "form_data_update",
   FORM_NEW_RECORD = "form_new_record",
   FORM_DATA_REMOVE = "form_data_remove",
   FORM_DATA_ADD = "form_data_add",
   AUTO_UPDATE = "auto_update"
}